requirements.txt
transformers
torch
faiss-cpu
wget
numpy
matplotlib
scikit-learn